nabu: sepolicy: basic initial policies for hwcontrol hal with service_nanager
* The custom aidl hal registers with service manager and is used by device settings app
with service_manager as a stub instance, So having a context of hwservice for the hal is
wrong as it produces a never allow as below:

08-15 00:14:38.816   644   644 E SELinux : avc:  denied  { find } for pid=3326 uid=1000 name=custom.hardware.hwcontrol.IHwControl/default scontext=u:r:devicesettings_app:s0 tcontext=u:object_r:default_android_service:s0 tclass=service_manager permissive=1

So fixup the policies and register the hal under correct context and sepolicies.
This patch also adds sepolicies for the various operatios done by hal and exposes 2 nodes

> touchfeature_device and vendor_sysfs_keyboard.

Earlier I labelled the keyboard sysfs node with sysfs_leds with genfs_contexts but with the
hal type that has coredomain we are not allowed to access content under sysfs unless label is
provided to them explicitly, So this time around they are labelled as vendor_sysfs_keyboard.

Fixes: dev-harsh1998/android_device_xiaomi_nabu@e169714

Signed-off-by: Harshit Jain <reach@harsh1998.dev>